+++ docs/resources/changelog.md
@@ -16,10 +16,13 @@ The most elegant solution I found was to use the [Web Animations API](https://de
 
 - 🚨 BREAKING: changed `left` and `right` placements to `start` and `end` in `sl-drawer`
 - 🚨 BREAKING: changed `left` and `right` placements to `start` and `end` in `sl-tab-group`
+- 🚨 BREAKING: removed `--hide-duration`, `--hide-timing-function`, `--show-duration`, and `--show-timing-function` custom properties from `sl-tooltip` (use the Animation Registry instead)
 - Added the Animation Registry
 - Fixed a bug where removing `sl-dropdown` from the DOM and adding it back destroyed the popover reference [#443](https://github.com/shoelace-style/shoelace/issues/443)
-- Updated animations for `sl-alert`, `sl-dialog`, `sl-drawer`, and `sl-dropdown` to use the Animation Registry instead of CSS transitions
+- Updated animations for `sl-alert`, `sl-dialog`, `sl-drawer`, `sl-dropdown`, and `sl-tooltip` to use the Animation Registry instead of CSS transitions
 - Improved a11y by respecting `prefers-reduced-motion` for all show/hide animations
+- Improved `--show-delay` and `--hide-delay` behavior in `sl-tooltip` so they only apply on hover
+- Removed the internal popover utility
 
 ## 2.0.0-beta.40